+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add tests
 - Add a CHANGELOG
+- Add support for Insta360 GO 2
+
+### Fixed
+- Only show progress for mounted disks
 
 ## [0.1.0] - 2023-05-22
 ### Added